Update watcher documentation
Hello, I'm not a Phoenix/Elixir expert, but I couldn't make the watcher work according to the doc. I was getting an error like this:

> ** (Protocol.UndefinedError) protocol Enumerable not implemented for {Tailwind, :install_and_run, [:default, ["--watch"]]} of type Tuple

So I poked around and changed the config to look more like node/webpack and this version works. I thought I'd create a PR for others take a look at this...