fallback to ipv4/ipv6 in the event of an unreachable host error
If there is a request failure and the reason implies that the error is
due to the usage being ipv4 or ipv6, the alternative is set and the
request retries. A variable is used to determine if a retry should be
attempted or not. If it is false, this means that a retry is already
being attempted, and the request should raise.

This is based on:

https://github.com/elixir-lang/elixir/blob/9c42348f6407fa315bae00acffe433e9417e2c2e/lib/mix/lib/mix/utils.ex#L687